Remove AddChainModal, ChainTile, ConnectWalletToolbox, and walletStore components to streamline the codebase. Update ConnectWallet to utilize publicClient directly for balance fetching and improve state management. Refactor related logic for better clarity and performance.
