--- toolbox/src/toolbox/components/ConnectWalletToolbox.tsx
@@ -1,309 +1,0 @@
-import { ConnectWallet } from "../../components/ConnectWallet";
-import { useL1ListStore, useViemChainStore } from "../toolboxStore";
-import { Button } from "../../components/Button";
-import * as Dialog from "@radix-ui/react-dialog";
-import { useEffect, useState } from "react";
-import { X } from "lucide-react";
-import { Input } from "../../components/Input";
-import { createPublicClient, http } from "viem";
-import { utils } from "@avalabs/avalanchejs";
-import { useWalletStore } from "../../lib/walletStore";
-import { Tabs } from "../../components/Tabs";
-import { getBlockchainInfo } from "../../coreViem/utils/glacier";
-import { Select } from "./Select";
-export const ConnectWalletToolbox = ({ children, required, chainRequired }: { children: React.ReactNode, required: boolean, chainRequired: boolean }) => {
-    const viemChain = useViemChainStore();
-
-    return (
-        <ConnectWallet required={required} extraElements={chainRequired ? <ChainSelector /> : null} >
-            {(chainRequired && !viemChain) ? (
-                <div className="p-4 border-2 border-gray-500 rounded-md mb-4">
-                    ⚠️ Please connect to an L1 chain before using this component.
-                </div>
-            ) : children}
-        </ConnectWallet>
-    );
-};
-
-const ChainSelector = () => {
-    const { l1List } = useL1ListStore();
-    const { walletChainId } = useWalletStore();
-
-    return (
-        <>
-            <div className="p-4 my-4 border-2 border-red-500 rounded-xl">Owen, put your chain selector here. This block is shown only if existing L1 is required.</div>
-            {/* <div className="flex items-center ">
-                {l1List.map(l1 => (
-                    <div key={l1.id} className={`cursor-pointer mr-4 mb-4 p-2 border rounded mb-1 ${walletChainId === l1.evmChainId ? "border-blue-500 border-2 font-bold" : "border-gray-300"}`} onClick={() => setLastSelectedL1(l1.id)}>
-                        <span>{l1.name || `Chain ${l1.id}`}</span>
-                    </div>
-                ))}
-
-
-            </div> 
-            <NewChainDialog />
-            */}
-        </>
-    );
-}
-
-// function NewChainDialog() {
-//     const { setLastSelectedL1, addL1 } = useL1ListStore();
-
-//     const [open, setOpen] = useState(false);
-//     const [rpcUrl, setRpcUrl] = useState("");
-//     const [chainId, setChainId] = useState("");
-//     const [evmChainId, setEvmChainId] = useState(0);
-//     const [coinName, setCoinName] = useState("COIN");
-//     const [isTestnet, setIsTestnet] = useState(true);
-//     const [rpcUrlError, setRpcUrlError] = useState("");
-//     const [name, setName] = useState("");
-//     const [subnetId, setSubnetId] = useState("");
-
-//     const FROM_RPC = "Enter RPC URL";
-//     const FROM_CORE_WALLET = "Load RPC URL from Core Wallet";
-
-//     const [activeTab, setActiveTab] = useState<typeof FROM_RPC | typeof FROM_CORE_WALLET>(FROM_RPC);
-
-//     useEffect(() => {
-//         async function fetchChainData() {
-//             setRpcUrlError("");
-//             setEvmChainId(0);
-//             setChainId("");
-//             setName("");
-
-//             if (!rpcUrl) {
-//                 return;
-//             }
-
-//             if (!rpcUrl.startsWith("https://")) {
-//                 setRpcUrlError("The RPC URL must start with https://");
-//                 return;
-//             }
-
-//             try {
-//                 const { ethereumChainId, avalancheChainId } = await fetchChainId(rpcUrl);
-//                 const subnetId = (await getBlockchainInfo(avalancheChainId)).subnetId;
-//                 const name = (await getBlockchainInfo(avalancheChainId)).blockchainName
-//                 setEvmChainId(ethereumChainId);
-//                 setChainId(avalancheChainId);
-//                 setSubnetId(subnetId);
-//                 setName(name);
-//             } catch (error) {
-//                 setRpcUrlError((error as Error)?.message || String(error));
-//             }
-//         }
-
-//         fetchChainData();
-//     }, [rpcUrl]);
-
-//     function addChain() {
-//         addL1({ id: chainId, name: name || `Chain ${evmChainId}`, rpcUrl: rpcUrl, evmChainId: evmChainId, coinName: coinName, isTestnet: isTestnet, subnetId: subnetId });
-//         setLastSelectedL1(chainId);
-//         setOpen(false);
-//     }
-
-//     return (
-//         <Dialog.Root open={open} onOpenChange={setOpen}>
-//             <Dialog.Trigger asChild>
-//                 <Button onClick={() => { }}>Add Chain</Button>
-//             </Dialog.Trigger>
-//             <Dialog.Portal>
-//                 <Dialog.Overlay className="fixed inset-0 bg-black/50 data-[state=open]:animate-overlayShow" />
-//                 <Dialog.Content className="fixed top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 bg-white p-6 rounded-xl shadow-lg focus:outline-none w-[90vw] max-w-md">
-//                     <Dialog.Title className="text-lg font-medium text-zinc-900 mb-4">
-//                         Add New Chain
-//                     </Dialog.Title>
-
-//                     <div className="space-y-4">
-//                         <Tabs
-//                             tabs={[FROM_RPC, FROM_CORE_WALLET]}
-//                             activeTab={activeTab}
-//                             setActiveTab={(tab) => setActiveTab(tab as typeof FROM_RPC | typeof FROM_CORE_WALLET)}
-//                             children={(activeTab) => {
-//                                 return <>
-//                                     {activeTab === FROM_RPC ?
-//                                         <ManualRPC rpcUrl={rpcUrl} setRpcUrl={setRpcUrl} /> :
-//                                         <ChainFromWallet
-//                                             rpcUrl={rpcUrl}
-//                                             setRpcUrl={setRpcUrl}
-//                                             setCoinName={setCoinName}
-//                                             setIsTestnet={setIsTestnet}
-//                                         />
-//                                     }
-//                                 </>
-//                             }}
-//                         />
-//                         {rpcUrlError && <p className="text-red-500">{rpcUrlError}</p>}
-
-//                         <Input
-//                             id="chainId"
-//                             label="Chain ID"
-//                             value={chainId}
-//                             disabled
-//                         />
-
-//                         <Input
-//                             id="evmChainId"
-//                             label="EVM Chain ID"
-//                             value={evmChainId}
-//                             disabled
-//                         />
-
-//                         <Input
-//                             id="name"
-//                             label="Chain Name"
-//                             value={name}
-//                             onChange={setName}
-//                         />
-
-//                         <Input
-//                             id="coinName"
-//                             label="Coin Name"
-//                             value={coinName}
-//                             onChange={setCoinName}
-//                         />
-
-//                         <Select
-//                             label="Is Testnet"
-//                             value={isTestnet ? "Yes" : "No"}
-//                             onChange={(value) => setIsTestnet(value === "Yes")}
-//                             options={[
-//                                 { label: "Yes", value: "Yes" },
-//                                 { label: "No", value: "No" },
-//                             ]}
-//                         />
-
-//                         <Button onClick={() => addChain()}>Add Chain</Button>
-//                     </div>
-
-//                     <Dialog.Close asChild>
-//                         <button
-//                             className="absolute top-4 right-4 inline-flex items-center justify-center rounded-full w-6 h-6 text-zinc-500 hover:text-zinc-700 focus:outline-none"
-//                             aria-label="Close"
-//                         >
-//                             <X size={16} />
-//                         </button>
-//                     </Dialog.Close>
-//                 </Dialog.Content>
-//             </Dialog.Portal>
-//         </Dialog.Root>
-//     )
-// }
-
-// function ManualRPC({ rpcUrl, setRpcUrl }: { rpcUrl: string, setRpcUrl: (url: string) => void }) {
-//     return (
-//         <Input
-//             id="rpcUrl"
-//             label="RPC Server URL"
-//             placeholder="https://..."
-//             value={rpcUrl}
-//             onChange={setRpcUrl}
-//         />
-//     )
-// }
-
-// function ChainFromWallet({
-//     rpcUrl,
-//     setRpcUrl,
-//     setCoinName,
-//     setIsTestnet
-// }: {
-//     rpcUrl: string,
-//     setRpcUrl: (url: string) => void,
-//     setCoinName: (name: string) => void,
-//     setIsTestnet: (isTestnet: boolean) => void
-// }) {
-//     const [anyChainId, setAnyChainId] = useState("");
-//     const [evmChainId, setEvmChainId] = useState(-1);
-//     const [localError, setLocalError] = useState("");
-//     const { walletChainId, coreWalletClient } = useWalletStore();
-
-//     //tries to extract evmChainId from anyChainId
-//     useEffect(() => {
-//         (async function () {
-//             setLocalError("");
-//             setEvmChainId(-1);
-//             if (!anyChainId) return
-
-//             console.log('anyChainId', anyChainId);
-
-//             try {
-//                 const parsed = parseInt(anyChainId, 10);
-//                 if (!isNaN(parsed)) {
-//                     setEvmChainId(parsed);
-//                     return
-//                 }
-//             } catch (e) {
-//                 console.log('error', e);
-//             }
-
-//             console.log('anyChainId', anyChainId);
-
-//             try {
-//                 utils.base58check.decode(anyChainId);//cause faulure before sending request
-//                 const chain = await getBlockchainInfo(anyChainId);
-//                 setEvmChainId(chain.evmChainId);
-//                 console.log('chain', chain);
-//             } catch (error) {
-//                 console.error("Failed to fetch chain info:", error);
-//                 setLocalError("The Chain ID doesn't seem to be an EVM Chain ID or Avalanche Chain ID")
-//             }
-//         })();
-//     }, [anyChainId]);
-
-//     async function switchToChain(evmChainId: number) {
-//         try {
-//             setLocalError("");
-
-//             const chainIdHex = `0x${parseInt(evmChainId.toString()).toString(16)}`;
-
-//             await window.avalanche!.request({
-//                 method: "wallet_switchEthereumChain",
-//                 params: [{ chainId: chainIdHex }],
-//             });
-//         } catch (error) {
-//             console.error("Failed to switch to chain:", error);
-//             setLocalError(`Failed to switch to chain, do you have chain ${evmChainId} added in your wallet? Original error: ${(error as Error)?.message || String(error)}`);
-//         }
-//     }
-
-//     async function fetchFromWallet() {
-//         try {
-//             setLocalError("");
-//             const evmInfo = await coreWalletClient.getEthereumChain()
-//             setRpcUrl(evmInfo.rpcUrls[0]);
-//             setCoinName(evmInfo.nativeCurrency.name);
-//             setIsTestnet(evmInfo.isTestnet);
-//         } catch (error) {
-//             console.error("Failed to fetch from wallet:", error);
-//             setLocalError("Failed to fetch from wallet, do you have chain ${evmChainId} added in your wallet?");
-//         }
-//     }
-
-//     return (
-//         <div className="space-y-4">
-//             <Input
-//                 id="chainId"
-//                 label="Any Chain ID - Avalanche or EVM"
-//                 value={anyChainId}
-//                 onChange={setAnyChainId}
-//                 error={localError}
-//             />
-
-//             {(walletChainId === evmChainId) && <Button onClick={fetchFromWallet} disabled={evmChainId === -1}>Fetch from wallet</Button>}
-//             {(walletChainId !== evmChainId) && <Button onClick={() => { switchToChain(evmChainId) }} disabled={evmChainId === -1}>Switch to chain {evmChainId}</Button>}
-
-//             {localError && <p className="text-red-500">{localError}</p>}
-
-//             <Input
-//                 id="rpcUrl"
-//                 label="RPC Server URL"
-//                 placeholder="https://..."
-//                 value={rpcUrl}
-//                 disabled
-//             />
-//         </div >
-//     )
-// }
-
